+++ test/amp/ReduceTest/ReduceTest.cpp
@@ -504,7 +504,7 @@ public:
         size_t size = GetParam( );
 
         std::generate(stdInput, stdInput + size, generateRandom<int>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltOutput[i] = stdInput[i];
@@ -542,7 +542,7 @@ public:
         size_t size = GetParam( );
 
         std::generate(stdInput, stdInput + size, generateRandom<float>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltOutput[i] = stdInput[i];
@@ -2068,7 +2068,7 @@ void testTBBDevicevector()
  //   bolt::amp::device_vector<int> tbbInput(aSize, 0);
     bolt::amp::device_vector<int> tbbInput(aSize);
 
-    for(int i=0; i<aSize; i++) {
+	for (size_t i = 0; i<aSize; i++) {
         stdInput[i] = i;
         tbbInput[i] = i;
     };
