+++ test/cl/InnerProductTest/InnerProductTest.cpp
@@ -342,7 +342,7 @@ public:
 
         std::generate(stdInput, stdInput + size, generateRandom<int>);
         std::generate(stdInput2, stdInput2 + size, generateRandom<int>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltInput2[i] = stdInput2[i];
@@ -380,7 +380,7 @@ public:
 
         std::generate(stdInput, stdInput + size, generateRandom<float>);
         std::generate(stdInput2, stdInput2 + size, generateRandom<float>);
-        for (int i = 0; i<size; i++)
+		for (size_t i = 0; i<size; i++)
         {
             boltInput[i] = stdInput[i];
             boltInput2[i] = stdInput2[i];
@@ -413,7 +413,7 @@ TEST( InnerProductStdVectWithInit, withIntWdInitWithStdPlusMinus)
     std::vector<int> boltInput (mySize);
     std::vector<int> boltInput2 (mySize);
 
-    for (int i = 0; i < mySize; ++i){
+	for (size_t i = 0; i < mySize; ++i){
         stdInput[i] = i;
         stdInput2[i] = i+1;
         boltInput[i] = stdInput[i];
@@ -440,7 +440,7 @@ TEST( CPUInnerProductStdVectWithInit, withIntWdInitWithStdPlusMinus)
     std::vector<int> boltInput (mySize);
     std::vector<int> boltInput2 (mySize);
 
-    for (int i = 0; i < mySize; ++i){
+	for (size_t i = 0; i < mySize; ++i){
         stdInput[i] = i;
         stdInput2[i] = i+1;
         boltInput[i] = stdInput[i];
@@ -471,7 +471,7 @@ TEST( MultiCoreInnerProductStdVectWithInit, withIntWdInitWithStdPlusMinus)
     std::vector<int> boltInput (mySize);
     std::vector<int> boltInput2 (mySize);
 
-    for (int i = 0; i < mySize; ++i){
+	for (size_t i = 0; i < mySize; ++i){
         stdInput[i] = i;
         stdInput2[i] = i+1;
         boltInput[i] = stdInput[i];
